asa
SBML2Escher:
* removed critical internal class from com.sun package that was used to
create an internal id for the map. Instead, the layout id is now used as
given in the SBML file.
* Since the debugger caused problems with the .forEach construct, I
changed some of them to regular for loops.
* I unified the forEach loop for reaction Glyphs, which significantly
speeds up the run time because before the entire list was traversed
several times.
* Refactoring: extracted several local variables from larger
class-method access chains to simplify debugging (e.g.,
textGlyph.getBoundingBox().getPosition().getX() deserves a variable for
the bounding box).
* Implemented a special case for common BiGG metaoblite ids: we now trim
the prefix "M_" if there is any.
* Improved the calculation of label coordinates for metabolite glyphs.
* Made code more easily readable by adding a few blanks here and there,
e.g., "cSs.size()-1" reads more easily like this "cSs.size() - 1", etc.
* Changed the id creation for multimarkers that often resulted in
deletion of existing multimarkers: often the id of speciesReferences was
not defined and the multimarker was then only named ".M" + some index,
e.g., ".M3" etc. Without a prefix it was not clear to which reaction
such a multimarker actually belonged, causing problems when rendering.
Now We are using the reaction glyph id and the speciesReferenceGlyph id
as prefix.
* Implemented a case distinction for reversing the direction of arcs
depending on whether they connect a reactant or a substrate to a
reaction midmarker.

* Improved error messages (removed line breaks in resource file and
specified type of expected argument, such as number)
* Minor improvements to indentation and formatting of code