Towards issue #51
The file GlycolysisLayout_small.sbml.xml seems entirely dysfunctional: It was missing a speciesGlyph and still backconversion (when ignoring validation) does not work, as Escher2Standard will enter into an infinite while-loop.
The program shows correct behaviour if --ignore-validation is not set (i.e. it does not even try to convert, because validation fails).